# packages/CoffeeClubApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CoffeeClub.Core.Api",
    "version": "v1"
  },
  "paths": {
...
    "/Order": {
      "get": {
        "tags": [
          "Order"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/Order/{orderId}/assign": {
      "post": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthProvider": {
        "enum": [
          "AzureActiveDirectory",
          "Google"
        ],
        "type": "string"
      },
      "CoffeeBean": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "roast": {
            "$ref": "#/components/schemas/Roast"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "inStock": {
            "type": "boolean"
          },
          "createdBy": {
            "$ref": "#/components/schemas/User"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CreateDrinkOrderDto": {
        "type": "object",
        "properties": {
          "coffeeBeanId": {
            "type": "string",
            "format": "uuid"
          },
          "drink": {
            "$ref": "#/components/schemas/Drink"
          },
          "milkType": {
            "$ref": "#/components/schemas/MilkType"
          },
          "isIced": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "drinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDrinkOrderDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Drink": {
        "enum": [
          "Filter",
          "Espresso",
          "Cappuccino",
          "Latte",
          "IcedLatte",
          "IcedCappuccino"
        ],
        "type": "string"
      },
      "DrinkOrderDto": {
        "required": [
          "coffeeBean",
          "drink"
        ],
        "type": "object",
        "properties": {
          "coffeeBean": {
            "$ref": "#/components/schemas/CoffeeBean"
          },
          "drink": {
            "$ref": "#/components/schemas/Drink"
          },
          "milkType": {
            "$ref": "#/components/schemas/MilkType"
          }
        },
        "additionalProperties": false
      },
...
      "MilkType": {
        "enum": [
          "Dairy",
          "Oat"
        ],
        "type": "string"
      },
      "OrderDto": {
        "required": [
          "drinks",
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "drinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DrinkOrderDto"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "additionalProperties": false
      },
      "OrderStatus": {
        "enum": [
          "Pending",
          "Received",
          "Assigned",
          "InProgress",
          "Ready"
        ],
        "type": "string"
      },
      "Roast": {
        "enum": [
          "Light",
          "Medium",
          "Dark"
        ],
        "type": "string"
      },
      "User": {
        "type": "object",
        "properties": {
          "authProvider": {
            "$ref": "#/components/schemas/AuthProvider"
          },
          "authId": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}